requirements.txt
qq-botpy>=1.2.0
schedule>=1.2.0
requests>=2.25.0

